player_client/requirements.txt
PyYAML
numpy
selenium
pyvirtualdisplay
# apt install xvfb xserver-xephyr tigervnc-standalone-server xfonts-base